Comic Store EXCELSIOR
EXCELSIOR Comic Store is a comprehensive SQL project that entails conceiving and implementing an intricate database. It comprises the systematic arrangement of distinct tables and columns, each populated with a variety of pertinent data. The project further incorporates the establishment of dynamic database views, along with the integration of procedural elements such as Stored Procedures and Triggers.